rename resource utils
