+++ src/index.ts
@@ -2,7 +2,7 @@ import * as fhirR4 from './FHIR-R4/classes/models-r4';
 import * as IfhirR4 from './FHIR-R4/interfaces/IModel';
 import * as fhirR3 from './FHIR-R3';
 import * as dstu2 from './FHIR-DSTU2';
-import { ResourceUtilities } from './library/ResourceUtils/ResourceUtilities';
+import { ResourceUtilities } from './library/ResourceUtils/ResourceUtils';
 import { BundleUtilities } from './library/BundleUtils/BundleUtils';
 export { fhirR4, fhirR3, IfhirR4, dstu2, ResourceUtilities, BundleUtilities };
