+++ src/library/ResourceUtils/ResourceUtils.ts
